Implement custom title search in document module
The document module in our blockchain application had basic CRUD
operations. However, it lacked the ability to perform a custom search
based on the title of the documents.

This commit introduces a new query, `search-title`, that allows users to
find all documents containing a specific word in their title. This
enhancement improves the usability of our application by providing more
flexible search options.

To implement this feature, we updated the Protobuf definitions and
regenerated the Protobuf files. We also adjusted the `SearchTitle`
function in `x/docchain/keeper/query_search_title.go` to return a slice
of `Document` objects that match the search criteria.

This is a significant step towards making our blockchain application more
user-friendly and versatile. Future work will focus on optimizing this
search function and introducing more complex queries.