Introduce 'category' field to 'document' type
Our blockchain application, DocChain, initially allowed documents to be
created with a 'title' and a 'body'. A new field, 'category', is introduced
to the 'document' type.

This change is consensus-breaking as it alters the transaction validation
rules. Nodes running the old version of the software will reject transactions
that include the 'category' field as these transactions do not conform to
the old rules. Conversely, nodes running the new version of the software
will reject transactions that do not include the 'category' field because
these transactions do not conform to the new rules.

In this commit, we have updated the relevant proto and keeper files to
include the 'category' field in the 'document' type. We have also updated
the README.md.